build: fix and overhaul build and ci
Targeting ESNext in the tsconfig was dangerous, that never
should have been set. Dropped the target down to ES2020.

Simplified the build with tsup, targeting Node v14 to align
with csrf-csrf. Flattened the build output.

Separated types from implementation.
Separated build and test ci stages.

This should not introduce any breaking changes.
