+++ package.json
@@ -3,27 +3,24 @@
   "version": "4.1.0",
   "description": "A utility package to help implement stateful CSRF protection using the Synchroniser Token Pattern in express.",
   "type": "module",
-  "main": "./lib/cjs/index.cjs",
+  "main": "./lib/index.cjs",
   "exports": {
     "require": {
       "types": "./lib/index.d.cts",
-      "default": "./lib/cjs/index.cjs"
+      "default": "./lib/index.cjs"
     },
     "import": {
       "types": "./lib/index.d.ts",
-      "import": "./lib/esm/index.js"
+      "import": "./lib/index.js"
     }
   },
   "types": "./lib/index.d.ts",
-  "files": ["lib/esm/index.js", "lib/cjs/index.cjs", "lib/index.d.ts", "lib/index.d.cts"],
+  "files": ["lib/index.js", "lib/index.cjs", "lib/index.d.ts", "lib/index.d.cts"],
   "scripts": {
     "clean": "rm -rf ./lib",
     "lint": "biome check .",
     "lint:fix": "biome check --write .",
-    "build:types": "tsc --declaration --outDir ./lib --emitDeclarationOnly",
-    "build:cjs": "tsc -p tsconfig.cjs.json && mv ./lib/cjs/index.js ./lib/cjs/index.cjs && cp ./lib/index.d.ts ./lib/index.d.cts",
-    "build:esm": "tsc -p tsconfig.json",
-    "build": "npm run build:types && npm run build:esm && npm run build:cjs",
+    "build": "tsup",
     "build:clean": "npm run clean && npm run build",
     "pack:clean": "rm -rf *.tgz && npm pack",
     "test": "vitest run",
@@ -37,12 +34,12 @@
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
     "@types/express": "^4.17.21",
-    "@types/express-session": "^1.17.10",
+    "@types/express-session": "^1.18.0",
     "@types/http-errors": "^2.0.4",
     "@types/node": "^18.15.8",
     "@vitest/coverage-istanbul": "3.1.2",
     "commit-and-tag-version": "12.5.0",
-    "ts-node": "^10.9.2",
+    "tsup": "8.4.0",
     "typescript": "^5.3.3",
     "vitest": "3.1.2"
   },
