+++ tsconfig.json
@@ -1,22 +1,22 @@
 {
   "compilerOptions": {
     "module": "ESNext",
-    "target": "ESNext",
+    "target": "ES2020",
     "esModuleInterop": true,
     "noImplicitAny": true,
     "moduleResolution": "node",
     "sourceMap": false,
-    "rootDir": "./src",
-    "outDir": "./lib/esm",
     "baseUrl": ".",
     "strictNullChecks": true,
     "isolatedModules": false,
     "declaration": false,
     "removeComments": true,
     "allowSyntheticDefaultImports": true,
+    "skipLibCheck": true,
     "paths": {
-      "*": ["node_modules/*"]
+      "@/*": ["./src/*"]
     }
   },
-  "include": ["./src/index.ts"]
+  "include": ["./src/**/*"],
+  "exclude": ["node_modules", "lib"]
 }
